Correctly load the Qt-native translations
Summary:
ecm_create_qm_loader() uses Q_COREAPP_STARTUP_FUNCTION to inject the
loader of the additiona translator. The macro does not seem to work
when the loader is linked in a separate static library (the code
which uses Qt translations is all inside stepcore).
So call the ECM macro (and then link the special loader) where
the Q(Core)Application is defined.
Many previously untranslated strings are not translated.
Some additional work is needed but it probably depends on the
special structure of stepcore.

Test Plan:
The code compiles, and the name of the objects (left pane) and the
properties (part of the right pane) are translated

Reviewers: #localization, #kde_edu, aacid

Reviewed By: #localization, aacid

Subscribers: aacid, #kde_edu, #localization

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D4793