Enable qalculate, which is not a Qt library
Qalculate has a separate Qt based interface (and a Gtk one)
but the library does not depend on Qt.
Enabling the library allows the program to parse the units
again when setting some values with a unit in the property editor
(for example 32 dm ± 10 cm is converted into 3.2 [m] ± 0.1 [m]).